R2: move provenance.py -> ir/hir/provenance.py
Provenance is an HIR concern (compiler-gen tracking for rules
synthesized by HIR rewrite passes — semi-join, head-constant, etc.).
The DSL only uses the USER_PROVENANCE sentinel constant; that import
moves over cleanly.

No behavior change. Updates 4 import sites:
  src/srdatalog/dsl.py
  src/srdatalog/ir/hir/emit.py
  src/srdatalog/ir/hir/rule_rewrite.py
  tests/test_semi_join.py
